backend/requirements.txt
Django==5.0.1
django-ninja==1.0.1
django-redis==5.4.0
djangorestframework-simplejwt==5.3.1
Pillow==10.2.0
python-dotenv==1.0.0
redis==5.0.1
django-tinymce==3.7.1
pymongo==4.6.1